move pke UpdateNodeGroupActivity next to other pke worflows
