+++ cmd/worker/main.go
@@ -383,7 +383,7 @@ func main() {
 
 		{
 			passwordSecrets := intpkeworkflowadapter.NewPasswordSecretStore(commonSecretStore)
-			err = registerPKEWorkflows(passwordSecrets, secret.Store)
+			err = registerPKEWorkflows(passwordSecrets)
 			if err != nil {
 				emperror.Panic(errors.WrapIf(err, "failed to register PKE workflows"))
 			}
